Implement ViewModels for each screen, refactor project file structue
