+++ app/src/main/java/com/example/kotlinapp/MainActivity.kt
@@ -2,6 +2,7 @@ package com.example.kotlinapp
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.kotlinapp.pokemon_list.PokemonListFragment
 
 class MainActivity : AppCompatActivity() {
