+++ app/src/main/java/com/example/kotlinapp/pokemon_database/FavoritePokemon.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinapp
+package com.example.kotlinapp.pokemon_database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
