+++ app/src/main/java/com/example/kotlinapp/pokemon_info/BaseStatAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinapp
+package com.example.kotlinapp.pokemon_info
 
 import android.content.res.ColorStateList
 import android.view.LayoutInflater
