+++ app/src/main/java/com/example/kotlinapp/pokemon_info/TypesAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinapp
+package com.example.kotlinapp.pokemon_info
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
