+++ app/src/main/java/com/example/kotlinapp/pokemon_list/PokemonItem.kt
@@ -1,4 +1,4 @@
-package com.example.kotlinapp
+package com.example.kotlinapp.pokemon_list
 
 class PokemonItem(
     val sprite: String,
